+++ android/flutter_rush_android.iml
@@ -75,5 +75,10 @@
     <orderEntry type="library" name="Gradle: junit:junit:4.12@jar" level="project" />
     <orderEntry type="library" name="Gradle: androidx.lifecycle:lifecycle-common:2.0.0@jar" level="project" />
     <orderEntry type="library" name="Gradle: androidx.webkit:webkit:1.0.0@aar" level="project" />
+    <orderEntry type="library" name="Gradle: com.umeng.umsdk:common:2.1.0@jar" level="project" />
+    <orderEntry type="library" name="Gradle: com.umeng.umsdk:push:6.0.1@aar" level="project" />
+    <orderEntry type="library" name="Gradle: com.umeng.umsdk:utdid:1.1.5.3@jar" level="project" />
+    <orderEntry type="library" name="Gradle: cn.jiguang.sdk:jcore:2.1.2@aar" level="project" />
+    <orderEntry type="library" name="Gradle: cn.jiguang.sdk:jpush:3.3.4@aar" level="project" />
   </component>
 </module>
